+++ lib/big_ml/batch_centroid.rb
@@ -1,19 +1,15 @@
 require 'big_ml/base'
 
 module BigML
-  class Evaluation < Base
-    EVALUATION_PROPERTIES = [
-      :category, :code, :combiner, :confidence_threshold, :created, :credits,
-      :dataset, :dataset_status, :description, :dev, :ensemble, :fields_map,
-      :locale, :max_rows, :missing_strategy, :model, :model_status,
-      :model_type, :name, :negative_class, :number_of_models, :ordering,
-      :out_of_bag, :positive_class, :private, :project, :range, :replacement,
-      :resource, :result, :rows, :sample_rate, :shared, :shared_hash,
-      :sharing_key, :size, :status, :subscription, :tags, :threshold, :type,
-      :updated
+  class BatchPrediction < Base
+    BATCH_PREDICTION_PROPERTIES = [
+      :category, :code, :created, :credits, :dataset, :dataset_status,
+      :description, :fields, :dataset, :model, :model_status, :name,
+      :objective_fields, :prediction, :prediction_path, :private, :resource,
+      :source, :source_status, :status, :tags, :updated
     ]
 
-    attr_reader *EVALUATION_PROPERTIES
+    attr_reader *BATCH_PREDICTION_PROPERTIES
 
     class << self
       def create(model_or_ensemble, dataset, options = {})
@@ -28,6 +24,16 @@ module BigML
         response = client.post("/#{resource_name}", {}, arguments.merge(options))
         self.new(response) if response.success?
       end
+
+      def download(id)
+        response = client.get("/#{resource_name}/#{id}/download")
+        response.body if response.success?
+      end
     end
+
+    def download
+      self.class.download(id)
+    end
+
   end
 end