+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.2.0
+-  BigML::Anomaly
+-  BigML::Cluster
+-  BigML::Centroid
+-  BigML::Score
+-  BigML::BatchCentroid
+-  BigML::BatchScore
+
+## 0.1.4
+- Update resources properties
+
+## 0.1.2 or 0.1.3
+- BigML::Ensemble
+- BigML::Evaluation
+
 ## 0.1.1
 - Casting method to_* added
 - Update and destroy methods in the instances